@@ -14,7 +14,6 @@
     "@types/jest": "^28.1.6",
     "@types/micromatch": "^4.0.1",
     "@types/node": "^18.17.15",
-    "@types/react": "^16.8.0",
     "@types/semver": "^7.1.0",
     "@yarnpkg/cli": "workspace:^",
     "@yarnpkg/core": "workspace:^",
@@ -24,16 +23,15 @@
     "@yarnpkg/sdks": "workspace:^",
     "clipanion": "^4.0.0-rc.2",
     "esbuild-wasm": "0.17.5",
-    "eslint": "^8.45.0",
+    "eslint": "^8.57.0",
     "jest": "^29.2.1",
     "pirates": "^4.0.5",
     "tslib": "^2.4.0",
-    "typescript": "5.4.1-rc"
+    "typescript": "5.5.0-beta"
   },
   "resolutions": {
     "ink": "patch:ink@npm%3A3.0.8#~/.yarn/patches/ink-npm-3.0.8-3a8005f59f.patch",
-    "yoga-layout-prebuilt": "patch:yoga-layout-prebuilt@1.10.0#./.yarn/patches/yoga-layout-prebuilt.patch",
-    "@docusaurus/mdx-loader": "patch:@docusaurus/mdx-loader@npm%3A2.4.3#~/.yarn/patches/%40docusaurus-mdx-loader-npm-2.4.3-0d95b4751a.patch"
+    "yoga-layout-prebuilt": "patch:yoga-layout-prebuilt@1.10.0#./.yarn/patches/yoga-layout-prebuilt.patch"
   },
   "dependenciesMeta": {
     "core-js": {
@@ -62,7 +60,7 @@
     "build:compile": "rm -rf \"$0\"/lib && mkdir -p \"$0\"/lib && rsync -a --include '*.d.ts' --exclude '*.ts' --exclude '*.tsx' \"$0\"/sources/ \"$0\"/lib/ && node scripts/compile \"$@\"",
     "test:lint": "eslint --max-warnings 0 \"./**/*.@(tsx|ts|js)\"",
     "test:unit": "jest",
-    "typecheck:all": "tsc --noEmit"
+    "typecheck:all": "tsc -p ."
   },
   "sherlock": {
     "requireList": [
@@ -79,11 +77,10 @@
   },
   "dependencies": {
     "@iarna/toml": "^2.2.5",
-    "@mdx-js/react": "^1.6.22",
     "@yarnpkg/types": "workspace:^",
     "chalk": "^3.0.0",
     "micromatch": "^4.0.2",
-    "react": "^17.0.2",
     "semver": "^7.1.2"
-  }
+  },
+  "packageManager": "yarn@4.2.1+sha256.15ce76682a8cd2090257b883cd69c637925b29573f9573e8403ec227d5ab6815"
 }
