@@ -39,18 +52,18 @@
     "os-browserify": "^0.3.0",
     "pako": "^2.0.4",
     "path-browserify": "^1.0.1",
-    "prism-react-renderer": "^1.3.5",
+    "prism-react-renderer": "^2.3.1",
     "process": "^0.11.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-instantsearch-dom": "^6.36.0",
     "react-instantsearch-hooks-web": "^6.36.0",
     "react-json-doc": "^2.3.4",
     "react-loading-skeleton": "^3.3.1",
     "react-modal": "^3.16.1",
     "react-query": "^3.39.3",
-    "react-select": "5.1.0",
-    "react-select-virtualized": "^5.3.0",
+    "react-select": "^5.8.0",
+    "react-select-virtualized": "^5.5.1",
     "react-tooltip": "^5.18.1",
     "react-virtualized": "^9.22.5",
     "react-window": "^1.8.9",
@@ -59,44 +72,58 @@
     "semver": "^7.1.2",
     "three": "^0.145.0",
     "tslib": "^2.4.0",
-    "unist-util-visit": "^2",
-    "unist-util-visit-parents": "^2",
+    "unist-util-visit-parents": "^6.0.1",
     "usehooks-ts": "^2.9.1",
     "webpack": "^5.88.2",
     "xterm": "^5.3.0",
     "xterm-addon-serialize": "^0.11.0",
-    "xterm-headless": "^5.3.0"
-  },
+    "xterm-headless": "^5.3.0   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.3",
-    "@docusaurus/theme-classic": "^2.4.3",
-    "@docusaurus/theme-common": "^2.4.3",
-    "@tsconfig/docusaurus": "^2.0.0",
+    "@docusaurus/module-type-aliases": "^3.2.1",
+    "@docusaurus/theme-classic": "^3.2.1",
+    "@docusaurus/theme-common": "^3.2.1",
+    "@docusaurus/tsconfig": "^3.2.1",
+    "@docusaurus/types": "^3.2.1",
     "@types/cross-spawn": "6.0.0",
     "@types/dompurify": "^3",
     "@types/git-url-parse": "^9.0.0",
     "@types/lodash": "^4.14.136",
     "@types/markdown-it": "^12.2.3",
     "@types/marked": "^5.0.0",
+    "@types/mdast": "^4.0.3",
+    "@types/node": "^18.19.15",
     "@types/pako": "^2",
-    "@types/react": "^16.8.0",
+    "@types/react": "^18.2.54",
     "@types/react-instantsearch-dom": "^6",
     "@types/react-modal": "^3",
     "@types/react-virtualized": "^9",
     "@types/react-window": "^1",
     "@types/resolve": "^1.20.0",
     "@types/semver": "^7.1.0",
+    
+<<<<<< snyk-fix-c5b4952db433f2657d54273a53f7d464
+    
     "@yarnpkg/core": "2.1.0",
+    
+=====
+    
+    "@types/unist": "^3.0.2",
+    "@yarnpkg/core": "workspace:^",
+    
+>>>>> master
+    
     "@yarnpkg/fslib": "workspace:^",
     "@yarnpkg/libzip": "workspace:^",
     "@yarnpkg/monorepo": "workspace:^",
     "@yarnpkg/shell": "workspace:^",
     "ansi-to-html": "^0.7.2",
-    "docusaurus-plugin-typedoc-api": "patch:docusaurus-plugin-typedoc-api@npm%3A3.0.1#~/.yarn/patches/docusaurus-plugin-typedoc-api-npm-3.0.1-7dd061feb2.patch",
+    "esast-util-from-js": "^2.0.1",
     "esbuild-loader": "^2.20.0",
     "fast-glob": "^3.2.2",
-    "typedoc": "^0.25.3",
-    "typescript": "5.4.1-rc"
+    "mdast-util-mdx-jsx": "^3.0.0",
+    "typedoc": "^0.25.7",
+    "typescript": "5.5.0-beta",
+    "unified": "^11.0.4"
   },
   "browserslist": {
     "production": [
