@@ -2,16 +2,17 @@
   "name": "@yarnpkg/eslint-config",
   "version": "1.0.0",
   "license": "BSD-2-Clause",
+  "type": "module",
   "exports": {
     ".": "./index.js",
     "./react": "./react.js"
   },
   "dependencies": {
-    "@rushstack/eslint-patch": "^1.3.2",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
+    "@typescript-eslint/eslint-plugin": "^7.8.0",
+    "@typescript-eslint/parser": "^7.8.0",
     "eslint-plugin-arca": "^0.16.0",
-    "eslint-plugin-react": "^7.33.0"
+    "eslint-plugin-react": "^7.34.1",
+    "globals": "^15.1.0"
   },
   "peerDependencies": {
     "eslint": "*",
@@ -31,7 +32,7 @@
     "node": ">=18.12.0"
   },
   "devDependencies": {
-    "eslint": "^8.45.0"
+    "eslint": "^8.57.0"
   },
   "stableVersion": "1.0.0"
 }
