@@ -1,6 +1,6 @@
 {
   "name": "@yarnpkg/builder",
-  "version": "4.0.0",
+  "version": "4.1.1",
   "license": "BSD-2-Clause",
   "bin": "./sources/boot-cli-dev.js",
   "exports": {
@@ -21,7 +21,7 @@
   "devDependencies": {
     "@types/semver": "^7.1.0",
     "@yarnpkg/monorepo": "workspace:^",
-    "typescript": "5.4.1-rc"
+    "typescript": "5.5.0-beta"
   },
   "scripts": {
     "postpack": "rm -rf lib",
@@ -47,5 +47,5 @@
   "engines": {
     "node": ">=18.12.0"
   },
-  "stableVersion": "4.0.0"
+  "stableVersion": "4.1.1"
 }
