@@ -27,9 +27,8 @@ const mte = generatePkgDriver({
       ? [projectFolder]
       : [];
 
-    const yarnBinary = process.env.TEST_FROM_SOURCES
-      ? require.resolve(`${__dirname}/../../../../../scripts/run-yarn.js`)
-      : require.resolve(`${__dirname}/../../../../yarnpkg-cli/bundles/yarn.js`);
+    const yarnBinary = process.env.TEST_BINARY
+      ?? require.resolve(`${__dirname}/../../../../yarnpkg-cli/bundles/yarn.js`);
 
     const res = await execFile(process.execPath, [...execArgv, yarnBinary, ...cwdArgs, command, ...args], {
       cwd: cwd || path,
