@@ -27,13 +27,98 @@ describe(`Plugins`, () => {
       );
 
       test(
-        `it should automatically enable automatic @types insertion when a tsconfig is detected`,
+        `it should automatically enable automatic @types insertion when a tsconfig is detected at the root of the project`,
         makeTemporaryEnv({}, async ({path, run, source}) => {
           await xfs.writeFilePromise(ppath.join(path, `tsconfig.json`), ``);
 
           await run(`add`, `is-number`);
 
           await expect(readManifest(path)).resolves.toMatchObject({
+            dependencies: {
+              [`is-number`]: `^2.0.0`,
+            },
+            devDependencies: {
+              [`@types/is-number`]: `^2`,
+            },
+          });
+        }),
+      );
+
+      test(
+        `it should automatically enable automatic @types insertion when a tsconfig is detected in the current workspace`,
+        makeTemporaryMonorepoEnv({
+          workspaces: [`packages/*`],
+        }, {[`packages/foo`]: {}}, async ({path, run, source}) => {
+          await xfs.writeFilePromise(ppath.join(path, `packages/foo/tsconfig.json`), ``);
+
+          await run(`add`, `is-number`, {
+            cwd: `${path}/packages/foo` as PortablePath,
+          });
+
+          await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.toMatchObject({
+            dependencies: {
+              [`is-number`]: `^2.0.0`,
+            },
+            devDependencies: {
+              [`@types/is-number`]: `^2`,
+            },
+          });
+        }),
+      );
+
+      test(
+        `it should automatically enable automatic @types insertion in the current workspace when tsEnableAutoTypes is set to true`,
+        makeTemporaryMonorepoEnv({
+          workspaces: [`packages/*`],
+        }, {[`packages/foo`]: {}}, async ({path, run, source}) => {
+          await run(`config`, `set`, `tsEnableAutoTypes`, `true`);
+
+          await run(`add`, `is-number`, {
+            cwd: `${path}/packages/foo` as PortablePath,
+          });
+
+          await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.toMatchObject({
+            dependencies: {
+              [`is-number`]: `^2.0.0`,
+            },
+            devDependencies: {
+              [`@types/is-number`]: `^2`,
+            },
+          });
+        }),
+      );
+
+      test(
+        `it should not automatically enable automatic @types insertion when a tsconfig is present in a sibling workspace`,
+        makeTemporaryMonorepoEnv({
+          workspaces: [`packages/*`],
+        }, {[`packages/foo`]: {}, [`packages/bar`]: {}}, async ({path, run, source}) => {
+          await xfs.writeFilePromise(ppath.join(path, `packages/foo/tsconfig.json`), ``);
+
+          await run(`add`, `is-number`, {
+            cwd: `${path}/packages/bar` as PortablePath,
+          });
+
+          await expect(readManifest(`${path}/packages/bar` as PortablePath)).resolves.toMatchObject({
+            dependencies: {
+              [`is-number`]: `^2.0.0`,
+            },
+          });
+        }),
+      );
+
+      test(
+        `it should automatically enable automatic @types insertion when a tsconfig is detected in the root project of the current workspace`,
+        makeTemporaryMonorepoEnv({
+          workspaces: [`packages/*`],
+        }, {[`packages/foo`]: {}}, async ({path, run, source}) => {
+          await xfs.writeFilePromise(ppath.join(path, `tsconfig.json`), ``);
+
+          await run(`add`, `is-number`, {
+            cwd: `${path}/packages/foo` as PortablePath,
+          });
+
+          await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.toMatchObject({
             dependencies: {
               [`is-number`]: `^2.0.0`,
             },
@@ -297,7 +382,24 @@ describe(`Plugins`, () => {
     describe(`Removing types`, () => {
       for (const type of Manifest.allDependencies) {
         test(
-          `it should automatically remove @types from ${type}`,
+          `it should not automatically remove @types from ${type} without tsconfig or tsEnableAutoTypes`,
+          makeTemporaryEnv(merge({
+            dependencies: {
+              [`is-number`]: `^1.0.0`,
+            },
+          }, {
+            [type]: {
+              [`@types/is-number`]: `1.0.0`,
+            },
+          }), {}, async ({path, run, source}) => {
+            await run(`remove`, `is-number`);
+
+            await expect(readManifest(path)).resolves.toHaveProperty(`${type}.@types/is-number`);
+          }),
+        );
+
+        test(
+          `it should automatically remove @types from ${type} with tsEnableAutoTypes set to true`,
           makeTemporaryEnv(merge({
             dependencies: {
               [`is-number`]: `^1.0.0`,
@@ -312,6 +414,102 @@ describe(`Plugins`, () => {
             await run(`remove`, `is-number`);
 
             await expect(readManifest(path)).resolves.not.toHaveProperty(`${type}.@types/is-number`);
+          }),
+        );
+
+        test(
+          `it should automatically remove @types from ${type} with tsconfig.json present`,
+          makeTemporaryEnv(merge({
+            dependencies: {
+              [`is-number`]: `^1.0.0`,
+            },
+          }, {
+            [type]: {
+              [`@types/is-number`]: `1.0.0`,
+            },
+          }), {}, async ({path, run, source}) => {
+            await xfs.writeFilePromise(ppath.join(path, `tsconfig.json`), ``);
+
+            await run(`remove`, `is-number`);
+
+            await expect(readManifest(path)).resolves.not.toHaveProperty(`${type}.@types/is-number`);
+          }),
+        );
+
+        test(
+          `it should not automatically remove @types ${type} from the current workspace without tsconfig.json present or tsEnableAutoTypes`,
+          makeTemporaryMonorepoEnv({
+            workspaces: [`packages/*`],
+          },
+          merge(
+            {[`packages/foo`]: {dependencies: {[`is-number`]: `^1.0.0`}}},
+            {[`packages/foo`]: {[type]: {[`@types/is-number`]: `1.0.0`}}},
+          ),
+          async ({path, run, source}) => {
+            await run(`remove`, `is-number`, {
+              cwd: `${path}/packages/foo` as PortablePath,
+            });
+
+            await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.toHaveProperty(`${type}.@types/is-number`);
+          }),
+        );
+
+        test(
+          `it should automatically remove @types ${type} from the current workspace when a tsconfig is detected in the root project of the current workspace`,
+          makeTemporaryMonorepoEnv({
+            workspaces: [`packages/*`],
+          },
+          merge(
+            {[`packages/foo`]: {dependencies: {[`is-number`]: `^1.0.0`}}},
+            {[`packages/foo`]: {[type]: {[`@types/is-number`]: `1.0.0`}}},
+          ),
+          async ({path, run, source}) => {
+            await xfs.writeFilePromise(ppath.join(path, `tsconfig.json`), ``);
+
+            await run(`remove`, `is-number`, {
+              cwd: `${path}/packages/foo` as PortablePath,
+            });
+
+            await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.not.toHaveProperty(`${type}.@types/is-number`);
+          }),
+        );
+
+        test(
+          `it should automatically remove @types ${type} from the current workspace with tsEnableAutoTypes set to true`,
+          makeTemporaryMonorepoEnv({
+            workspaces: [`packages/*`],
+          },
+          merge(
+            {[`packages/foo`]: {dependencies: {[`is-number`]: `^1.0.0`}}},
+            {[`packages/foo`]: {[type]: {[`@types/is-number`]: `1.0.0`}}},
+          ),
+          async ({path, run, source}) => {
+            await run(`config`, `set`, `tsEnableAutoTypes`, `true`);
+            await run(`remove`, `is-number`, {
+              cwd: `${path}/packages/foo` as PortablePath,
+            });
+
+            await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.not.toHaveProperty(`${type}.@types/is-number`);
+          }),
+        );
+
+        test(
+          `it should automatically remove @types ${type} from the current workspace with tsconfig.json present`,
+          makeTemporaryMonorepoEnv({
+            workspaces: [`packages/*`],
+          },
+          merge(
+            {[`packages/foo`]: {dependencies: {[`is-number`]: `^1.0.0`}}},
+            {[`packages/foo`]: {[type]: {[`@types/is-number`]: `1.0.0`}}},
+          ),
+          async ({path, run, source}) => {
+            await xfs.writeFilePromise(ppath.join(path, `packages/foo/tsconfig.json`), ``);
+
+            await run(`remove`, `is-number`, {
+              cwd: `${path}/packages/foo` as PortablePath,
+            });
+
+            await expect(readManifest(`${path}/packages/foo` as PortablePath)).resolves.not.toHaveProperty(`${type}.@types/is-number`);
           }),
         );
       }
