@@ -26,8 +26,10 @@ const afterWorkspaceDependencyAddition = async (
   const {project} = workspace;
   const {configuration} = project;
 
-  const tsEnableAutoTypes = configuration.get(`tsEnableAutoTypes`)
-    ?? xfs.existsSync(ppath.join(project.cwd, `tsconfig.json`));
+  const tsEnableAutoTypes = configuration.get(`tsEnableAutoTypes`) ?? (
+    xfs.existsSync(ppath.join(workspace.cwd, `tsconfig.json`))
+      || xfs.existsSync(ppath.join(project.cwd, `tsconfig.json`))
+  );
 
   if (!tsEnableAutoTypes)
     return;
@@ -118,8 +120,10 @@ const afterWorkspaceDependencyRemoval = async (
   const {project} = workspace;
   const {configuration} = project;
 
-  const tsEnableAutoTypes = configuration.get(`tsEnableAutoTypes`)
-    ?? xfs.existsSync(ppath.join(project.cwd, `tsconfig.json`));
+  const tsEnableAutoTypes = configuration.get(`tsEnableAutoTypes`) ?? (
+    xfs.existsSync(ppath.join(workspace.cwd, `tsconfig.json`))
+      || xfs.existsSync(ppath.join(project.cwd, `tsconfig.json`))
+  );
 
   if (!tsEnableAutoTypes)
     return;
