@@ -17,7 +17,7 @@ export default class VersionApplyCommand extends BaseCommand {
     details: `
       This command will apply the deferred version changes and remove their definitions from the repository.
 
-      Note that if \`--prerelease\` is set, the given prerelease identifier (by default \`rc.%d\`) will be used on all new versions and the version definitions will be kept as-is.
+      Note that if \`--prerelease\` is set, the given prerelease identifier (by default \`rc.%n\`) will be used on all new versions and the version definitions will be kept as-is.
 
       By default only the current workspace will be bumped, but you can configure this behavior by using one of:
 
