@@ -1,9 +1,9 @@
-import {VirtualFS, npath}               from '@yarnpkg/fslib';
-import fs                               from 'fs';
-import {fileURLToPath, pathToFileURL}   from 'url';
+import {VirtualFS, npath}                                                                            from '@yarnpkg/fslib';
+import fs                                                                                            from 'fs';
+import {fileURLToPath, pathToFileURL}                                                                from 'url';
 
-import {WATCH_MODE_MESSAGE_USES_ARRAYS} from '../loaderFlags';
-import * as loaderUtils                 from '../loaderUtils';
+import {SUPPORTS_IMPORT_ATTRIBUTES, SUPPORTS_IMPORT_ATTRIBUTES_ONLY, WATCH_MODE_MESSAGE_USES_ARRAYS} from '../loaderFlags';
+import * as loaderUtils                                                                              from '../loaderUtils';
 
 // The default `load` doesn't support reading from zip files
 export async function load(
@@ -11,6 +11,9 @@ export async function load(
   context: {
     format: string | null | undefined;
     importAssertions?: {
+      type?: 'json';
+    };
+    importAttributes?: {
       type?: 'json';
     };
   },
@@ -26,10 +29,21 @@ export async function load(
   if (!format)
     return nextLoad(urlString, context, nextLoad);
 
-  if (format === `json` && context.importAssertions?.type !== `json`) {
-    const err = new TypeError(`[ERR_IMPORT_ASSERTION_TYPE_MISSING]: Module "${urlString}" needs an import assertion of type "json"`) as TypeError & { code: string };
-    err.code = `ERR_IMPORT_ASSERTION_TYPE_MISSING`;
-    throw err;
+  if (format === `json`) {
+    if (SUPPORTS_IMPORT_ATTRIBUTES_ONLY) {
+      if (context.importAttributes?.type !== `json`) {
+        const err = new TypeError(`[ERR_IMPORT_ATTRIBUTE_MISSING]: Module "${urlString}" needs an import attribute of "type: json"`) as TypeError & { code: string };
+        err.code = `ERR_IMPORT_ATTRIBUTE_MISSING`;
+        throw err;
+      }
+    } else {
+      const type = `importAttributes` in context ? context.importAttributes?.type : context.importAssertions?.type;
+      if (type !== `json`) {
+        const err = new TypeError(`[ERR_IMPORT_ASSERTION_TYPE_MISSING]: Module "${urlString}" needs an import ${SUPPORTS_IMPORT_ATTRIBUTES ? `attribute` : `assertion`} of type "json"`) as TypeError & { code: string };
+        err.code = `ERR_IMPORT_ASSERTION_TYPE_MISSING`;
+        throw err;
+      }
+    }
   }
 
   // https://github.com/nodejs/node/pull/44366/files#diff-f6796082f599554ec3a29c47cf026cb24fc5104884f2632e472c05fe622d778bR477-R479
