@@ -118,22 +118,25 @@ const TEMPLATE = (relPnpApiPath: PortablePath, module: string, {setupEnv = false
   `#!/usr/bin/env node\n`,
   `\n`,
   `const {existsSync} = require(\`fs\`);\n`,
-  `const {createRequire} = require(\`module\`);\n`,
+  `const {createRequire, register} = require(\`module\`);\n`,
   `const {resolve} = require(\`path\`);\n`,
+  `const {pathToFileURL} = require(\`url\`);\n`,
   `\n`,
   `const relPnpApiPath = ${JSON.stringify(npath.fromPortablePath(relPnpApiPath))};\n`,
   `\n`,
   `const absPnpApiPath = resolve(__dirname, relPnpApiPath);\n`,
   `const absRequire = createRequire(absPnpApiPath);\n`,
   `\n`,
-  ...(wrapModule ? [
-    `const moduleWrapper = ${wrapModule.trim().replace(/^ {4}/gm, ``)}\n`,
-    `\n`,
-  ] : []),
+  `const absPnpLoaderPath = resolve(absPnpApiPath, \`../.pnp.loader.mjs\`);\n`,
+  `const isPnpLoaderEnabled = existsSync(absPnpLoaderPath);\n`,
+  `\n`,
   `if (existsSync(absPnpApiPath)) {\n`,
   `  if (!process.versions.pnp) {\n`,
   `    // Setup the environment to be able to require ${module}\n`,
   `    require(absPnpApiPath).setup();\n`,
+  `    if (isPnpLoaderEnabled && register) {\n`,
+  `      register(pathToFileURL(absPnpLoaderPath));\n`,
+  `    }\n`,
   `  }\n`,
   ...(setupEnv ? [
     `\n`,
@@ -164,6 +167,10 @@ const TEMPLATE = (relPnpApiPath: PortablePath, module: string, {setupEnv = false
   ] : []),
   `}\n`,
   `\n`,
+  ...(wrapModule ? [
+    `const moduleWrapper = ${wrapModule.trim().replace(/^ {4}/gm, ``)}\n`,
+    `\n`,
+  ] : []),
   `// Defer to the real ${module} your application uses\n`,
   wrapModule ? `module.exports = moduleWrapper(absRequire(\`${module}\`));\n` : `module.exports = absRequire(\`${module}\`);\n`,
 ].join(``);
